+++ pkg/util/limiter/memory_consumption_test.go
@@ -29,7 +29,7 @@ func TestFromContextWithFallback(t *testing.T) {
 	t.Run("exists", func(t *testing.T) {
 		ctx := context.Background()
 		existing := NewMemoryConsumptionTracker(0, nil, "")
-		require.NoError(t, existing.IncreaseMemoryConsumption(uint64(512)))
+		require.NoError(t, existing.IncreaseMemoryConsumption(uint64(512), IngesterChunks))
 
 		ctx = context.WithValue(ctx, memoryConsumptionTracker, existing)
 		stored := MemoryTrackerFromContextWithFallback(ctx)
@@ -41,7 +41,7 @@ func TestFromContextWithFallback(t *testing.T) {
 func TestAddToContext(t *testing.T) {
 	ctx := context.Background()
 	existing := NewMemoryConsumptionTracker(0, nil, "")
-	require.NoError(t, existing.IncreaseMemoryConsumption(uint64(512)))
+	require.NoError(t, existing.IncreaseMemoryConsumption(uint64(512), IngesterChunks))
 
 	ctx = AddMemoryTrackerToContext(ctx, existing)
 	stored := ctx.Value(memoryConsumptionTracker).(*MemoryConsumptionTracker)
@@ -53,34 +53,35 @@ func TestMemoryConsumptionTracker_Unlimited(t *testing.T) {
 	reg, metric := createRejectedMetric()
 	tracker := NewMemoryConsumptionTracker(0, metric, "foo + bar")
 
-	require.NoError(t, tracker.IncreaseMemoryConsumption(128))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(128, IngesterChunks))
 	require.Equal(t, uint64(128), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(128), tracker.PeakEstimatedMemoryConsumptionBytes())
 
 	// Add some more memory consumption. The current and peak stats should be updated.
-	require.NoError(t, tracker.IncreaseMemoryConsumption(2))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(2, StoreGatewayChunks))
 	require.Equal(t, uint64(130), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(130), tracker.PeakEstimatedMemoryConsumptionBytes())
 
 	// Reduce memory consumption. The current consumption should be updated, but the peak should be unchanged.
-	tracker.DecreaseMemoryConsumption(128)
+	tracker.DecreaseMemoryConsumption(128, IngesterChunks)
 	require.Equal(t, uint64(2), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(130), tracker.PeakEstimatedMemoryConsumptionBytes())
 
 	// Add some more memory consumption that doesn't take us over the previous peak.
-	require.NoError(t, tracker.IncreaseMemoryConsumption(8))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(8, FPointSlices))
 	require.Equal(t, uint64(10), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(130), tracker.PeakEstimatedMemoryConsumptionBytes())
 
 	// Add some more memory consumption that takes us over the previous peak.
-	require.NoError(t, tracker.IncreaseMemoryConsumption(121))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(121, HPointSlices))
 	require.Equal(t, uint64(131), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(131), tracker.PeakEstimatedMemoryConsumptionBytes())
 
 	assertRejectedQueriesCount(t, reg, 0)
 
 	// Test reducing memory consumption to a negative value panics
-	require.PanicsWithValue(t, `Estimated memory consumption of this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(150) })
+	require.PanicsWithValue(t, `Estimated memory consumption of all instances of []promql.FPoint in this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(9, FPointSlices) })
+	require.PanicsWithValue(t, `Estimated memory consumption of all instances of []promql.HPoint in this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(130, HPointSlices) })
 }
 
 func TestMemoryConsumptionTracker_Limited(t *testing.T) {
@@ -88,51 +89,52 @@ func TestMemoryConsumptionTracker_Limited(t *testing.T) {
 	tracker := NewMemoryConsumptionTracker(11, metric, "foo + bar")
 
 	// Add some memory consumption beneath the limit.
-	require.NoError(t, tracker.IncreaseMemoryConsumption(8))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(8, IngesterChunks))
 	require.Equal(t, uint64(8), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(8), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 0)
 
 	// Add some more memory consumption beneath the limit.
-	require.NoError(t, tracker.IncreaseMemoryConsumption(1))
+	require.NoError(t, tracker.IncreaseMemoryConsumption(1, StoreGatewayChunks))
 	require.Equal(t, uint64(9), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(9), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 0)
 
 	// Reduce memory consumption.
-	tracker.DecreaseMemoryConsumption(1)
+	tracker.DecreaseMemoryConsumption(1, StoreGatewayChunks)
 	require.Equal(t, uint64(8), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(9), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 0)
 
 	// Try to add some more memory consumption where we would go over the limit.
 	const expectedError = "the query exceeded the maximum allowed estimated amount of memory consumed by a single query (limit: 11 bytes) (err-mimir-max-estimated-memory-consumption-per-query)"
-	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(4), expectedError)
+	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(4, StoreGatewayChunks), expectedError)
 	require.Equal(t, uint64(8), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(9), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 1)
 
 	// Make sure we don't increment the rejection count a second time for the same query.
-	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(4), expectedError)
+	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(4, IngesterChunks), expectedError)
 	require.Equal(t, uint64(8), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(9), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 1)
 
 	// Keep adding more memory consumption up to the limit to make sure the failed increases weren't counted.
 	for i := 0; i < 3; i++ {
-		require.NoError(t, tracker.IncreaseMemoryConsumption(1))
+		require.NoError(t, tracker.IncreaseMemoryConsumption(1, FPointSlices))
 		require.Equal(t, uint64(9+i), tracker.CurrentEstimatedMemoryConsumptionBytes())
 		require.Equal(t, uint64(9+i), tracker.PeakEstimatedMemoryConsumptionBytes())
 	}
 
 	// Try to add some more memory consumption when we're already at the limit.
-	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(1), expectedError)
+	require.ErrorContains(t, tracker.IncreaseMemoryConsumption(1, FPointSlices), expectedError)
 	require.Equal(t, uint64(11), tracker.CurrentEstimatedMemoryConsumptionBytes())
 	require.Equal(t, uint64(11), tracker.PeakEstimatedMemoryConsumptionBytes())
 	assertRejectedQueriesCount(t, reg, 1)
 
 	// Test reducing memory consumption to a negative value panics
-	require.PanicsWithValue(t, `Estimated memory consumption of this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(150) })
+	require.PanicsWithValue(t, `Estimated memory consumption of all instances of []promql.FPoint in this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(150, FPointSlices) })
+	require.PanicsWithValue(t, `Estimated memory consumption of all instances of []promql.HPoint in this query is negative. This indicates something has been returned to a pool more than once, which is a bug. The affected query is: foo + bar`, func() { tracker.DecreaseMemoryConsumption(150, HPointSlices) })
 }
 
 func assertRejectedQueriesCount(t *testing.T, reg *prometheus.Registry, expectedRejectionCount int) {
@@ -157,13 +159,14 @@ func BenchmarkMemoryConsumptionTracker(b *testing.B) {
 	// limit since this should be rare. Instead, we want to benchmark having to check the
 	// limit.
 	const memoryLimit = 1024 * 1024 * 1024
+	const source = IngesterChunks
 
 	b.Run("no limits single threaded", func(b *testing.B) {
 		l := NewMemoryConsumptionTracker(0, nil, "")
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			_ = l.IncreaseMemoryConsumption(uint64(b.N))
-			l.DecreaseMemoryConsumption(uint64(b.N))
+			_ = l.IncreaseMemoryConsumption(uint64(b.N), source)
+			l.DecreaseMemoryConsumption(uint64(b.N), source)
 		}
 
 		require.Equal(b, uint64(0), l.CurrentEstimatedMemoryConsumptionBytes())
@@ -177,8 +180,8 @@ func BenchmarkMemoryConsumptionTracker(b *testing.B) {
 		l := NewMemoryConsumptionTracker(memoryLimit, counter, "")
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			if err := l.IncreaseMemoryConsumption(uint64(b.N)); err == nil {
-				l.DecreaseMemoryConsumption(uint64(b.N))
+			if err := l.IncreaseMemoryConsumption(uint64(b.N), source); err == nil {
+				l.DecreaseMemoryConsumption(uint64(b.N), source)
 			}
 		}
 
@@ -195,15 +198,15 @@ func BenchmarkMemoryConsumptionTracker(b *testing.B) {
 			defer wg.Done()
 
 			for run.Load() {
-				_ = l.IncreaseMemoryConsumption(1024)
-				l.DecreaseMemoryConsumption(1024)
+				_ = l.IncreaseMemoryConsumption(1024, source)
+				l.DecreaseMemoryConsumption(1024, source)
 			}
 		}()
 
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			_ = l.IncreaseMemoryConsumption(uint64(b.N))
-			l.DecreaseMemoryConsumption(uint64(b.N))
+			_ = l.IncreaseMemoryConsumption(uint64(b.N), source)
+			l.DecreaseMemoryConsumption(uint64(b.N), source)
 		}
 
 		run.Store(false)
@@ -226,16 +229,16 @@ func BenchmarkMemoryConsumptionTracker(b *testing.B) {
 			defer wg.Done()
 
 			for run.Load() {
-				if err := l.IncreaseMemoryConsumption(1024); err == nil {
-					l.DecreaseMemoryConsumption(1024)
+				if err := l.IncreaseMemoryConsumption(1024, source); err == nil {
+					l.DecreaseMemoryConsumption(1024, source)
 				}
 			}
 		}()
 
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			if err := l.IncreaseMemoryConsumption(uint64(b.N)); err == nil {
-				l.DecreaseMemoryConsumption(uint64(b.N))
+			if err := l.IncreaseMemoryConsumption(uint64(b.N), source); err == nil {
+				l.DecreaseMemoryConsumption(uint64(b.N), source)
 			}
 		}
 
@@ -244,5 +247,10 @@ func BenchmarkMemoryConsumptionTracker(b *testing.B) {
 
 		require.Equal(b, uint64(0), l.CurrentEstimatedMemoryConsumptionBytes())
 	})
+}
 
+func TestMemoryConsumptionSourceNames(t *testing.T) {
+	for i := range memoryConsumptionSourceCount {
+		require.NotEqual(t, unknownMemorySource, i.String(), "source %d should have a String() representation", i)
+	}
 }